Refactor: Move Voices resource download to ViewModel
Moves the orchestration of extracting resource IDs and downloading associated resources
for the news list from `VoicesFragment` to `VoicesViewModel`. This moves the I/O-heavy
operations off the UI layer.
