+++ app/src/main/java/org/ole/planet/myplanet/ui/voices/VoicesFragment.kt
@@ -34,7 +34,6 @@ import org.ole.planet.myplanet.services.VoicesLabelManager
 import org.ole.planet.myplanet.ui.chat.ChatDetailFragment
 import org.ole.planet.myplanet.ui.components.FragmentNavigator
 import org.ole.planet.myplanet.utils.FileUtils
-import org.ole.planet.myplanet.utils.JsonUtils.getString
 import org.ole.planet.myplanet.utils.KeyboardUtils.setupUI
 import org.ole.planet.myplanet.utils.Utilities
 import org.ole.planet.myplanet.utils.collectLatestWhenStarted
@@ -174,7 +173,7 @@ class VoicesFragment : BaseVoicesFragment() {
         val sortedList = sortNews(list)
         if (binding.rvNews.adapter == null) {
             changeLayoutManager(resources.configuration.orientation, binding.rvNews)
-            downloadResourcesForNews(sortedList)
+            voicesViewModel.downloadReferencedResources(sortedList)
             setupVoicesAdapter(sortedList.filterNotNull())
         } else {
             (binding.rvNews.adapter as? VoicesAdapter)?.submitList(sortedList.filterNotNull()) {
@@ -185,25 +184,6 @@ class VoicesFragment : BaseVoicesFragment() {
             }
         }
         showNoData(binding.tvMessage, sortedList.filterNotNull().size, currentEmptyStateSource)
-    }
-
-    private fun downloadResourcesForNews(list: List<News?>) {
-        val resourceIds = mutableSetOf<String>()
-        list.forEach { news ->
-            if ((news?.imagesArray?.size() ?: 0) > 0) {
-                val ob = news?.imagesArray?.get(0)?.asJsonObject
-                val resourceId = getString("resourceId", ob?.asJsonObject)
-                if (!resourceId.isNullOrBlank()) {
-                    resourceIds.add(resourceId)
-                }
-            }
-        }
-        viewLifecycleOwner.lifecycleScope.launch {
-            if (resourceIds.isNotEmpty()) {
-                val libraries = resourcesRepository.getLibraryItemsByIds(resourceIds)
-                resourcesRepository.downloadResources(libraries)
-            }
-        }
     }
 
     private fun sortNews(list: List<News?>): List<News?> {
